requirements.txt
python-dotenv==1.0.0
duckdb==0.9.0
